requirements.txt
discord.py>=2.3.2,<3.0.0
python-dotenv>=1.0.0